+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Change are listed in reverse chronological order (newest to oldest).
 
 <span class="changelog">
 
+###### [ 1.0.175 ] - 2025/02/23
+
+  * Updated `SpotifyClient` methods to utilize the Spotify Web Player elevated access token (if configured) if a Spotify Player method is accessed.  This should prevent issues with Sonos devices.  Note that the [Spotify Web Player Authentication Setup](https://github.com/thlucas1/homeassistantcomponent_spotifyplus/wiki/Device-Configuration-Options#spotify-web-player-authentication-setup) must be configured to use this functionality; if not configured, then the Sonos Controller (e.g. SoCo) API is used to control the Sonos device.  The following methods were updated: `PlayerMediaPause`, `PlayerMediaResume`, `PlayerMediaSeek`, `PlayerMediaSkipNext`, `PlayerMediaSkipPrevious`, `PlayerSetRepeatMode`, `PlayerSetShuffleMode`, `PlayerSetVolume`, `PlayerTransferPlayback`, `AddPlayerQueueItems`.
+
 ###### [ 1.0.174 ] - 2025/02/22
 
   * Updated `SpotifyClient` methods to restore functions that were previously deprecated by the Spotify development team.  Note that the [Spotify Web Player Authentication Setup](https://github.com/thlucas1/homeassistantcomponent_spotifyplus/wiki/Device-Configuration-Options#spotify-web-player-authentication-setup) must be enabled to use these specific functions; if not enabled, the functions will still work but won't return Spotify owned items and data.  The following methods have functionality restored: `GetPlaylistItems`.